consolidate new tests
